requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0

# Deep Learning
torch>=1.9.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.61.0
pyyaml>=5.4.0

# Optional: CUDA support for PyTorch
# Install appropriate version based on your CUDA version
# torch-cuda>=1.9.0 

# Added from the code block
jupyter>=1.0.0
notebook>=6.4.0
black>=21.6b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0
pre-commit>=2.13.0 